Move auth to middlewares
